fix: resolve mypy type errors in target_client_builder and redteam CLI
- target_client_builder.py: fix `best` tuple annotation from
  `tuple[int, ...]` to `tuple[float, ...]` — score is float
  (node.confidence * 10)
- redteam.py: update both `_run_redteam` return type annotations from
  `tuple[list, dict, list, str]` to `tuple[list, dict, list, str, list[str]]`
  to match the 5-tuple actually returned (added config_notes in a prior commit)

+++ nuguard/common/target_client_builder.py
@@ -122,7 +122,7 @@ def _discover_login_endpoint(sbom: "AiSbomDocument") -> tuple[str, str, str] | N
     except Exception:
         return None
 
-    best: tuple[int, str, str, str] | None = None  # (score, path, user_field, pass_field)
+    best: tuple[float, str, str, str] | None = None  # (score, path, user_field, pass_field)
     for node in sbom.nodes:
         if node.component_type != NodeType.API_ENDPOINT:
             continue
